fix: make node switching work with custom URLs
ServiceNow stores an encoded identifier for the node to connect to in
the cookie `BIGipServerpool_<instance name>`.

We were calculating the cookie name by pulling instance name from the
URL, however we cannot do this with a custom URL.

This meant we were creating a new cookie, rather than editing the one
that drives node connection.

This change makes setActiveNode pull the cookie name from the existing
cookie rather than calculating it, ensuring we set the value of the
right cookie.

Fixes #136. More details can be found there.