fix: generate correct node BIG IP encoded port
ServiceNow stores an encoded identifier for the node to connect to in
the cookie `BIGipServerpool_<instance name>`.

We have to calculate this value in order to set the cookie to direct
us to the desired node. This calculation was failing for the encoded
port section of the BIG IP because we were pulling node stats from
an incorrect location. I suspect an upgrade changed the location
they are stored in but cannot confirm that.

This commit makes us pull the node stats xml from the correct
location, so that we can parse the port and encode it.

More details can be found in issue #138