Fixes scripts not loading in studio
CodeMirror, the library used by pre-Polaris script fields cannot be instantiated if the element it is being inserted into is not visible.

When opening many records with script fields in Studio, only the current tab is visible, so many records won't load their script fields.

ServiceNow's implementation checks once if it's visible and then gives up. We now have InteractionObserver a relatively performant way to check when an element becomes visible.

We can find all the script fields on load and check if they have backing GlideEditor objects. If not, we add an observation handler to load the field once the user returns.

Once that occurs, if the script field is successfully loaded we can disconnect the handler.

Since this needs to be injected into every frame, I felt this was the best implementation but I'm open to other ways.