Slash command to clear breakpoints
This command clears all breakpoints for a user and updates the active tab's code editor with the removal.

Supports both CodeMirror and Monaco.

We fetch all the breakpoints from the table api, but for our removals we use the /js/debugger endpoint used by the frontend because breakpoints appear to be session cached so removing breakpoints via the table api takes a while to reflect on the current session.

We use the batch api to make the calls, its significantly faster than individual requests as we only have to pass ACL's and authentication once.

Once we remove all the breakpoints, we attempt to refresh the editor's via their loadPoints methods.

The CodeMirror version is more involved as loadPoints for some reason doesn't remove the line background added by a breakpoint, so we have to do that manually.

I tried to make the batch request functions mildly reusable, but it may have just overcomplicated it.

I also added reusable role check methods and iframe location methods, since I needed them myself. If they pass muster there is a decent amount of places they can fit into.